fix: change all id types to bigint (long) in table creation

+++ src/main/scala/db/CreateTables.scala
@@ -17,15 +17,15 @@ def createTables(xa: Transactor) = {
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${Ingredients.table}(
-          ${Ingredients.table.id} SERIAL PRIMARY KEY,
+          ${Ingredients.table.id} BIGSERIAL PRIMARY KEY,
           ${Ingredients.table.name} VARCHAR(255) NOT NULL
         );
       """,
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${Storages.table}(
-          ${Storages.table.id} SERIAL PRIMARY KEY,
-          ${Storages.table.ownerId} INT NOT NULL,
+          ${Storages.table.id} BIGSERIAL PRIMARY KEY,
+          ${Storages.table.ownerId} BIGINT NOT NULL,
           ${Storages.table.name} VARCHAR(255) NOT NULL,
           FOREIGN KEY (${Storages.table.ownerId}) REFERENCES ${Users.table}(${Users.table.id}) ON DELETE CASCADE
         );
@@ -33,8 +33,8 @@ def createTables(xa: Transactor) = {
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${StorageMembers.table}(
-          ${StorageMembers.table.storageId} INT NOT NULL,
-          ${StorageMembers.table.memberId} INT NOT NULL,
+          ${StorageMembers.table.storageId} BIGINT NOT NULL,
+          ${StorageMembers.table.memberId} BIGINT NOT NULL,
           PRIMARY KEY (${StorageMembers.table.storageId}, ${StorageMembers.table.memberId}),
           FOREIGN KEY (${StorageMembers.table.storageId}) REFERENCES ${Storages.table}(${Storages.table.id}) ON DELETE CASCADE,
           FOREIGN KEY (${StorageMembers.table.memberId}) REFERENCES ${Users.table}(${Users.table.id}) ON DELETE CASCADE
@@ -43,8 +43,8 @@ def createTables(xa: Transactor) = {
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${StorageIngredients.table}(
-          ${StorageIngredients.table.storageId} INT NOT NULL,
-          ${StorageIngredients.table.ingredientId} INT NOT NULL,
+          ${StorageIngredients.table.storageId} BIGINT NOT NULL,
+          ${StorageIngredients.table.ingredientId} BIGINT NOT NULL,
           PRIMARY KEY (${StorageIngredients.table.storageId}, ${StorageIngredients.table.ingredientId}),
           FOREIGN KEY (${StorageIngredients.table.storageId}) REFERENCES ${Storages.table}(${Storages.table.id}) ON DELETE CASCADE,
           FOREIGN KEY (${StorageIngredients.table.ingredientId}) REFERENCES ${Ingredients.table}(${Ingredients.table.id}) ON DELETE CASCADE
@@ -53,7 +53,7 @@ def createTables(xa: Transactor) = {
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${Recipes.table}(
-          ${Recipes.table.id} SERIAL PRIMARY KEY,
+          ${Recipes.table.id} BIGSERIAL PRIMARY KEY,
           ${Recipes.table.name} VARCHAR(255) NOT NULL,
           ${Recipes.table.sourceLink} VARCHAR(128) NOT NULL
         );
@@ -61,8 +61,8 @@ def createTables(xa: Transactor) = {
 
       sql"""
         CREATE TABLE IF NOT EXISTS ${RecipeIngredients.table}(
-          ${RecipeIngredients.table.recipeId} INT NOT NULL,
-          ${RecipeIngredients.table.ingredientId} INT NOT NULL,
+          ${RecipeIngredients.table.recipeId} BIGINT NOT NULL,
+          ${RecipeIngredients.table.ingredientId} BIGINT NOT NULL,
           PRIMARY KEY (${RecipeIngredients.table.recipeId}, ${StorageIngredients.table.ingredientId}),
           FOREIGN KEY (${RecipeIngredients.table.recipeId}) REFERENCES ${Recipes.table}(${Recipes.table.id}) ON DELETE CASCADE,
           FOREIGN KEY (${RecipeIngredients.table.ingredientId}) REFERENCES ${Ingredients.table}(${Ingredients.table.id}) ON DELETE CASCADE
